requirements.txt
Django==2.0.1
django-cors-headers==2.2.0
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.8.2
mysqlclient==1.3.12
requests==2.18.4
uWSGI==2.0.17
